requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.9.2
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.10.10=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.4.1=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.6.2.post1=pypi_0
attrs=24.2.0=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.9.24=hca03da5_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
dataclasses-json=0.6.7=pypi_0
expat=2.6.3=h313beb8_0
faiss-cpu=1.9.0=pypi_0
frozenlist=1.5.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.6=pypi_0
httpx=0.27.2=pypi_0
httpx-sse=0.4.0=pypi_0
idna=3.10=pypi_0
jinja2=3.1.4=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
langchain=0.3.7=pypi_0
langchain-community=0.3.5=pypi_0
langchain-core=0.3.15=pypi_0
langchain-text-splitters=0.3.2=pypi_0
langsmith=0.1.142=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
marshmallow=3.23.1=pypi_0
mdurl=0.1.2=pypi_0
multidict=6.1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
narwhals=1.13.3=pypi_0
ncurses=6.4=h313beb8_0
numpy=1.26.4=pypi_0
openssl=3.0.15=h80987f9_0
orjson=3.10.11=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pillow=11.0.0=pypi_0
pip=24.2=py312hca03da5_0
propcache=0.2.0=pypi_0
protobuf=5.28.3=pypi_0
pyarrow=18.0.0=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pydantic-settings=2.6.1=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.18.0=pypi_0
pypdf=5.1.0=pypi_0
python=3.12.7=h99e199e_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h1a28f6b_0
referencing=0.35.1=pypi_0
requests=2.32.3=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.21.0=pypi_0
setuptools=75.1.0=py312hca03da5_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.35=pypi_0
sqlite=3.45.3=h80987f9_0
streamlit=1.40.0=pypi_0
tenacity=9.0.0=pypi_0
tk=8.6.14=h6ba3021_0
toml=0.10.2=pypi_0
tornado=6.4.1=pypi_0
typing-extensions=4.12.2=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
wheel=0.44.0=py312hca03da5_0
xz=5.4.6=h80987f9_1
yarl=1.17.1=pypi_0
zlib=1.2.13=h18a0788_1
